Minimal baseline working high-level fetch interface to Maude. Backend data source is set via fetch.from_cxc() and fetch.from_maude().
What's missing so far from the Maude interface:
- No caching of the results from a Maude fetch
- Retrieval of a set of MSID's are, for now, performed one at a
time. When fetching from Maude this incurs a considerable penalty in
round-trip overhead.
- Values retrieved from Maude are all considered 'good'. No filtering
on 'bad' values is applied.
- No asynchronous notification of service interruption when connected
to the MAUDE server. This may leave fetch in an inconsistent state if
MAUDE went down while we were connected to it. A temporary remedy is
to simply reset the backend to CXC via fetch.from_cxc()